WiP: Doxygen parser produces parse tree.
(WiP, unstable)

Mixing section commands with math commands as is done in hipSparse
is breaking the parser.

This forced us to do a rewrite and a change to the user interface,
where the transform_* methods have been replaced by a parse
method that emits a parse tree for a given doxygen comment.

TODO:

Integrate with _codegen.cython codegeneration backend.