fix(i*gen/tree*): improve handling of anon typedef record/enum
Background
----------

Target languages:

We consider target languages like Cython that do not allow to have a struct
and a typedef carry the same name, i.e. the following is not allowed
`struct name` and `typedef struct name name`.
Further, types cannot be anonymous.

Libclang:

For

`typedef struct /* no name / same name / different name */ { ... } typedef_name;`

(or union/enum) on the top-level of the translation unit,
libclang generates an anonymous outer struct (or union/enum)
plus a typedef that contains the tree struct/union/enum subtree again.
This is only performed on the top-level and not for nested typedef'ed
elaborated types.

Implementation aspects:

When the underlying typedef type is encountered, it is not clear
that it stems from a typedef construct. Hence, we traverse the
inner types and put a `tree.Node` per found type (including the parent) into
a type lookup table.

Nodes of nested types and their parents are linked via parent and children fields.
Updating these fields and the type lookup table becomes non-trivial when
a type is removed.

Type removal happened before this commit but was only ever working for trivial scenarios.
We decided now to keep the underlying typedef type also in case of an
anonymous type and assigning it a name a posteriori.
This ensures that the lookup table and the type hiearchies remain unchanged.