+  patch for numba installation
Allows to patch an existing numba (with cuda)
installation in the currently active conda environment.

Conda lib folder and Python version are determind
automatically by the script.

Background:

It is easier to installa a recent numba installation
and then patch in the numba/roc folder than
building numba from scratch.