examples(2_Advanced): + device linking example
In this example, we link a device code file that contains
a device kernel "__global__ void print_tid()" that has an unresolved call
to a function "__device__ void foo()" in its body with a second
device code file that contains the definition of device function "foo".

To make this work, both snippets need to be compiled with
the ``-fgpu-rdc`` option and the compilation results needs to
be added as `HIPRTC_JIT_INPUT_LLVM_BITCODE` type input to the link object.