examples(2_*): + hiprtc_linking_llvm_ir.py
In this example, we link a device code snippet that contains
a HIP C++ device kernel ``__global__ void scale(float[],float)`` that has an unresolved call
to a function ``__device__ void scale_op(float[],float)`` in its body with a
LLVM IR snippet that contains the definition of that device function ``scale_op``.
(Note that we could have also used LLVM BC (bitcode) instead of human-readable LLVM IR.)

To make this work, the HIP C++ snippet needs to be compiled with
the ``-fgpu-rdc`` option and all compilation results needs to
be added as ``HIPRTC_JIT_INPUT_LLVM_BITCODE`` type input to the HIPRTC link object.

Note that the LLVM IR in this example is target dependent.
Therefore, this example can currently only be run with ``gfx90a`` (MI200 series).
The example will quit with an error if other architectures are used.