hipify(test_cuda_array_slicing): partially
Two issues:

Many tests require JIT to work, skipped
for now.

Non-contiguous memory copies are failing in this test.
For some reason, strides are not taken into account by the
slicing when creating a new memory buffer for carrying the sliced out elements.
This would be fine if the strides were taken into account
when doing host<->device copies but this seems also not to be the case.